+++ app/controllers/api/collections_controller.rb
@@ -5,7 +5,7 @@ class Api::CollectionsController < ApplicationController
     before_action :set_user, only: [:user_collections]
 
     def index
-        render json: current_user.collections, include: [:cards]
+        render json: current_user.collections.order(:id), include: [:cards]
     end
 
     def all_collections
@@ -38,6 +38,8 @@ class Api::CollectionsController < ApplicationController
     end
 
     def destroy
+        # map over the user to check for collection id is in liked_collections, remove it and update user's liked_collections.
+        # User.all
         render json: @collection.destroy
     end
