+++ client/src/components/CollectionLike.js
@@ -7,37 +7,47 @@ import FavoriteIcon from '@mui/icons-material/Favorite';
 
 const CollectionLike = (props) => {
   const auth = useContext(AuthContext);
-  const [liked, setLiked] = useState(false);
+  const [liked, setLiked] = useState(null);
   const [like, setLike] = useState(props.collection.likes)
-  console.log(props.collection.likes)
-  console.log(auth)
-
-  useEffect(() => {
-    isLiked();
-  }, [])
-
-  const isLiked = () => {
-    if (auth.liked_collections.includes(props.collection.id)) {
-      setLiked(true)
-    }
+  const [likedArray, setLikedArray] = useState([]);
+  
+  useEffect(()=>{
+    getUser();
+  },[]);
+  
+  const getUser = async () => {
+    let res = await axios.get(`/api/users/${auth.id}`)
+    setLikedArray(res.data.liked_collections)
+    setLiked(res.data.liked_collections.includes(props.collection.id) ? true : false)
   }
 
-  const handleLike = () => {
+  const handleLike = async () => {
+    // like: liked ? likes -1 : likes +1
     if (!liked) {
-      setLiked(true)
+      setLiked(true);
+      let likingArray = [...likedArray, props.collection.id]
       let liked = like + 1
+      let updateUserLikes = await axios.put(`/api/users/liked_collections`, { liked_collections: likingArray })
+      let updateCollectionLikes = await axios.put(`/api/collections/${props.collection.id}`, { likes: liked })
       setLike(liked)
     }
     else {
-      setLiked(false)
+      setLiked(false);
+      let unlikedArray = likedArray.filter((c) => props.collection.id !== c)
       let unliked = like - 1
+      let removeUserLikes = await axios.put(`/api/users/liked_collections`, { liked_collections: unlikedArray })
+      let removeCollectionLikes = await axios.put(`/api/collections/${props.collection.id}`, { likes: unliked })
       setLike(unliked)
     }
   }
 
-
   return (
-    <button onClick = {()=>{handleLike()}}>{liked === false ? <FavoriteBorderIcon/> : <FavoriteIcon />}{like}</button>
+    <button
+      onClick={() => { handleLike() }}
+    >
+      {liked === false ? <FavoriteBorderIcon /> : <FavoriteIcon />}
+      {like}
+    </button>
   );
 }
